requirements.txt
requests
flask
flask_cors
mfrc522
RPi.GPIO